rename `analysis_name` -> `name_analysis`
