+++ paddle/fluid/pir/utils/CMakeLists.txt
@@ -4,6 +4,6 @@ cc_library(
   DEPS op_dialect op_dialect_vjp pir)
 
 cc_library(
-  pir_analysis_name
-  SRCS analysis_name.cc
+  pir_name_analysis
+  SRCS name_analysis.cc
   DEPS op_dialect op_dialect_vjp pir)
