+++ paddle/fluid/pybind/CMakeLists.txt
@@ -534,7 +534,7 @@ if(WITH_PYTHON)
     list(APPEND PYBIND_DEPS eager_tensor_operants)
     list(APPEND PYBIND_DEPS pybind_util)
     # pir utils
-    list(APPEND PYBIND_DEPS pir_analysis_name)
+    list(APPEND PYBIND_DEPS pir_name_analysis)
   endif()
 
   # On Linux, cc_library(paddle SHARED ..) will generate the libpaddle.so,
